Implement Patch Documents
This commit creates a class that represents a JSON document. The three methods provided will give us enough flexibility to implement all patch operations.

Unfortunately I haven't found a way to avoid adding a type annotation to account for the fact that when a pointer is created as part of #split_pointer, it will have at least one reference token (because all public methods have separate handling for empty pointers).