Implement Patch Operations
This commit adds classes to perform all of the patch operations, and then uses them to perform the patch on the target document.

Having finally implemented the `JSON::Patch.call` behavior, we can add back the IETF tests to confirm that our JSON::Patch library conforms to the [JSON Patch Specification Doc](https://datatracker.ietf.org/doc/html/rfc6902)

With this commit we're introducing a dependency on "duplicate" - ruby doesn't have a nice way of deeply duplicating an object and while most of this is written from scratch, implementing deep duplication seems unnecessary.